requirements.txt
genshinstats @ git+https://github.com/nitolar/genshinstats
python-dotenv
win10toast
pyttsx3